- split web into multiple files
- ManualResetEvent for RequestRefresh and WaitForRefresh
- double buffering _rgb
- passing _imageBuffer to JavaScript as pinned pointer (to not copy bytes on every frame)
