+++ GB.WASM/GB.WASM.csproj
@@ -2,31 +2,32 @@
   <PropertyGroup>
     <TargetFramework>net9.0</TargetFramework>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-	<WasmEnableThreads>true</WasmEnableThreads>
+    <WasmEnableThreads>true</WasmEnableThreads>
+
     <RunAOTCompilation>true</RunAOTCompilation>
-		<WasmEnableSIMD>true</WasmEnableSIMD>
+    <WasmEnableSIMD>true</WasmEnableSIMD>
+    <!--
+    <WasmNativeDebugSymbols>true</WasmNativeDebugSymbols>
+    <WasmNativeStrip>false</WasmNativeStrip>
+    -->
   </PropertyGroup>
 
-	<ItemGroup>
-        <PackageReference Include="Microsoft.NET.WebAssembly.Threading" Version="9.0.3" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\GB.Core\GB.Core.csproj" />
-	</ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\GB.Core\GB.Core.csproj" />
+  </ItemGroup>
 
   <ItemGroup>
     <Compile Update="TestResources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>TestResources.resx</DependentUpon>
+    <DesignTime>True</DesignTime>
+    <AutoGen>True</AutoGen>
+    <DependentUpon>TestResources.resx</DependentUpon>
     </Compile>
   </ItemGroup>
 
   <ItemGroup>
     <EmbeddedResource Update="TestResources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>TestResources.Designer.cs</LastGenOutput>
+    <Generator>ResXFileCodeGenerator</Generator>
+    <LastGenOutput>TestResources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
 </Project>
